requirements.txt
# GUARDIAN ML — Python Dependencies
# Install: pip install -r requirements.txt

# ── Core Framework ────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# ── ML & Data Science ─────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.2
scipy==1.13.0

# ── Visualization ─────────────────────────────────────────
plotly==5.22.0
matplotlib==3.8.4
seaborn==0.13.2
kaleido==0.2.1

# ── Geospatial ────────────────────────────────────────────
geopandas==0.14.4
shapely==2.0.4
pyproj==3.6.1
folium==0.16.0

# ── Configuration & Utilities ─────────────────────────────
pyyaml==6.0.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Logging & Monitoring ──────────────────────────────────
loguru==0.7.2
rich==13.7.1

# ── File Handling ─────────────────────────────────────────
openpyxl==3.1.2
aiofiles==23.2.1

# ── Testing ───────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
